+++ modyn/selector/internal/selector_strategies/presampling_strategies/label_balanced_presampling_strategy.py
@@ -4,6 +4,4 @@ from modyn.selector.internal.selector_strategies.presampling_strategies import A
 
 class LabelBalancedPresamplingStrategy(AbstractBalancedPresamplingStrategy):
     def __init__(self, presampling_config: dict, modyn_config: dict, pipeline_id: int):
-        super().__init__(presampling_config, modyn_config, pipeline_id)
-
-        self.balanced_column = SelectorStateMetadata.label
+        super().__init__(presampling_config, modyn_config, pipeline_id, SelectorStateMetadata.label)
